dasdas

+++ database/migrations/2025_05_22_064658_create_prices_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('price_in');
             $table->integer('price_out');
-            $table->integer('type');
+            $table->bigInteger('type');
             $table->string('url', 500);
             $table->timestamp('published_at')->nullable();
             $table->timestamps();
